refactor: 디렉토리 구조 수정 및 401/404 페이지 추가
