+++ src/components/product/ProductGrid.jsx
@@ -1,4 +1,4 @@
-import ProductItem from "./ProductItem";
+import ProductItem from "./ProductItem.jsx";
 
 export default function ProductGrid({
                                       products,
