+++ src/pages/seller/ProductRegister.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import ProductImageUploader from "../components/ProductImageUploader";
+import ProductImageUploader from "../../components/product/ProductImageUploader.jsx";
 import axios from "axios";
 import { v4 as uuidv4 } from "uuid";
